# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Artichaut API",
        "description": "API documentation for Hotel Artichaut API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get one news by id",
                "operationId": "60ce16d414a9c3e71989ed2f67cdb9a0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the news",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Service not found"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Update a news article by id",
                "operationId": "74e290b2773173f71c79a30460a14ba7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the news article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "title",
                                    "short_description",
                                    "description",
                                    "display_order",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name to group with other languages",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "The title of the news article",
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "description": "Short description of the news article",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Full description of the news article",
                                        "type": "string"
                                    },
                                    "display_order": {
                                        "description": "The desired disaly order the items should be",
                                        "type": "integer"
                                    },
                                    "language_id": {
                                        "description": "The ID of the language",
                                        "type": "integer"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "description": "An array of image files",
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete a news article by id",
                "operationId": "654c643d6d36c35debc9e6259cc2e95d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the news article",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "News not found"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
        "/api/news/lang-{lang}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "operationId": "aaec8649c26e426b81ca4265bd45485f",
                "parameters": [
                    {
                        "name": "lang",
                        "in": "path",
                        "description": "The lang desired",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            }
        },
        "/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "operationId": "a73927f8b55e83ba817c6f5378e0b08f",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "500": {
                        "description": "An error occured"
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Add a news article",
                "operationId": "ad8caa89d9ea11c22d751612547bad6c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "title",
                                    "short_description",
                                    "description",
                                    "display_order",
                                    "language_id"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name to group with other languages",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "The title of the news article",
                                        "type": "string"
                                    },
                                    "short_description": {
                                        "description": "Short description of the news article",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Full description of the news article",
                                        "type": "string"
                                    },
                                    "display_order": {
                                        "description": "The desired disaly order the items should be",
                                        "type": "integer"
                                    },
                                    "language_id": {
                                        "description": "The ID of the language",
                                        "type": "integer"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "description": "An array of image files",
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Validation failed"
                    },
                    "500": {
                        "description": "An error occurred"
                    }
                }
            }
        },
...
    },
...
}